rename to ResourceSelectorContent
