+++ apps/studio/src/components/ResourceSelector/ResourceSelectorContent.tsx
